Update harvest script
