--- brownie/scripts/realloc.py
@@ -1,3 +1,0 @@
-from aerodrome_harvest import *
-def main():
-    harvest_and_swap()
